+++ lib/util/logger.js
@@ -1,0 +1,5 @@
+function log (msg) {
+  console.error(msg)
+}
+
+module.exports = log
